+++ src/transformer.ts
@@ -2,6 +2,7 @@ import { mappings } from './mappings';
 import { generator_options } from './types/generator';
 import * as esprima from 'esprima';
 import { generator } from './generator';
+import { check } from 'prettier';
 
 /**
  * This will replace code in AST pre-rebuilding
@@ -47,8 +48,12 @@ export const transformer = (ast: any, options: generator_options) => {
   };
 
   const replaceReturnedExpression = (x: any) => {
+    if (x?.type === 'Identifier') {
+      return x;
+    }
+
     let esp_initialising_vars = getInstanceInitialising(ast);
-    let device_variable: string = x.callee.object.object.name;
+    let device_variable: string = x?.callee?.object?.object?.name;
 
     let device_init: any = esp_initialising_vars.find(
       (x: any) => x.name === device_variable,
@@ -81,11 +86,79 @@ export const transformer = (ast: any, options: generator_options) => {
     return ast_res;
   };
 
-  const replaceExpression = (x: any) => {
+  const replaceExpression = (x: any): any => {
     let esp_initialising_vars = getInstanceInitialising(ast);
     let device_variable: string;
 
-    if (x.expression.type === 'ConditionalExpression') {
+    switch (x?.type) {
+      case 'IfStatement': {
+        return replaceIfStatement(x);
+      }
+      case 'ClassDeclaration':
+        return replaceClass(x);
+      case 'FunctionDeclaration':
+      case 'WhileStatement':
+      case 'ForStatement':
+      case 'ForInStatement':
+      case 'DoWhileStatement': {
+        return replaceLoopStatement(x);
+      }
+      case 'SwitchStatement':
+        return replaceSwitchStatement(x);
+      case 'ContinueStatement':
+      case 'BreakStatement':
+        return x;
+    }
+
+    if (x?.type === 'VariableDeclaration') {
+      if (x.declarations[0].init?.type === 'FunctionExpression') {
+        x.declarations[0].init.body.body = x.declarations[0].init.body.body.map(
+          (y: any) => replaceExpression(y),
+        );
+      }
+
+      if (x.declarations[0].init?.type === 'AwaitExpression') {
+        return x;
+      }
+
+      if (x.declarations[0].init?.type === 'ObjectExpression') {
+        x.declarations[0].init.properties =
+          x.declarations[0].init.properties.map((y: any) => {
+            if (y.value.type === 'FunctionExpression') {
+              y.value = replaceLoopStatement(y.value);
+            } else if (y.value.type === 'CallExpression') {
+              y.value = replaceReturnedExpression(y.value);
+            } else if (x.value.type === 'ArrowFunctionExpression') {
+              y.value.body = replaceReturnedExpression(y.value.body);
+            }
+            return y;
+          });
+
+        return x;
+      }
+    }
+
+    if (x?.type === 'ReturnStatement') {
+      if (x.argument.type === 'Literal') {
+        return x;
+      }
+
+      x.argument = replaceReturnedExpression(x.argument);
+
+      return x;
+    }
+
+    if (x?.expression.type === 'LogicalExpression') {
+      if (x.expression.right.type !== 'Literal') {
+        x.expression.right = replaceReturnedExpression(x.expression.right);
+      }
+      if (x.expression.left.type !== 'Literal') {
+        x.expression.left = replaceReturnedExpression(x.expression.left);
+      }
+      return x;
+    }
+
+    if (x?.expression.type === 'ConditionalExpression') {
       x.expression.consequent = replaceReturnedExpression(
         x.expression.consequent,
       );
@@ -95,21 +168,26 @@ export const transformer = (ast: any, options: generator_options) => {
       return x;
     }
 
-    if (x.expression.type === 'AssignmentExpression') {
+    if (x?.expression.type === 'AssignmentExpression') {
       return x;
     }
-    if (x.expression.callee.object.type === 'MemberExpression') {
+
+    if (x?.expression?.callee?.object.type === 'MemberExpression') {
       device_variable = x.expression.callee.object.object.name;
-    } else if (x.expression.callee.object.type === 'Identifier') {
+    } else if (x?.expression?.callee?.object?.type === 'Identifier') {
       device_variable = x.expression.callee.object.name;
     } else {
       device_variable = '';
     }
 
+    if (x?.expression?.object?.type === 'ThisExpression') {
+      return x;
+    }
+
     if (
       !esp_initialising_vars.map((x: any) => x.name).includes(device_variable)
     ) {
-      if (x.expression.arguments instanceof Array) {
+      if (x?.expression.arguments instanceof Array) {
         x.expression.arguments = x.expression.arguments.map((y: any) => {
           if (y.hasOwnProperty('value')) {
             return y.value;
@@ -177,11 +255,26 @@ export const transformer = (ast: any, options: generator_options) => {
             ast.declarations[0].init.properties.map((x: any) => {
               if (x.value.type === 'FunctionExpression') {
                 x.value = replaceLoopStatement(x.value);
-              } else {
+              } else if (x.value.type === 'CallExpression') {
                 x.value = replaceReturnedExpression(x.value);
+              } else if (x.value.type === 'ArrowFunctionExpression') {
+                x.value.body = replaceReturnedExpression(x.value.body);
               }
               return x;
             });
+        }
+
+        if (ast.declarations[0].init?.type === 'ArrowFunctionExpression') {
+          if (ast.declarations[0].init.body.body instanceof Array) {
+            ast.declarations[0].init.body.body =
+              ast.declarations[0].init.body.body.map((x: any) =>
+                replaceExpression(x),
+              );
+          } else {
+            ast.declarations[0].init.body = replaceReturnedExpression(
+              ast.declarations[0].init.body,
+            );
+          }
         }
 
         if (ast.declarations[0].init.hasOwnProperty('callee')) {
@@ -252,13 +345,27 @@ export const transformer = (ast: any, options: generator_options) => {
     return class_copy;
   };
 
+  const replaceTryCatch = (x: any) => {
+    x.block.body = x.block.body.map((y: any) => {
+      return replaceExpression(y);
+    });
+    x.handler.body.body = x.handler.body.body.map((y: any) => {
+      return replaceExpression(y);
+    });
+    return x;
+  };
+
   const getExpressions = (ast: any): any => {
     let ast_copy: any = { ...ast };
+
     ast_copy.body = ast.body
       .map((x: any) => {
         switch (x.type) {
           case 'ExpressionStatement': {
             return replaceExpression(x);
+          }
+          case 'TryStatement': {
+            return replaceTryCatch(x);
           }
           case 'IfStatement': {
             return replaceIfStatement(x);
@@ -269,6 +376,7 @@ export const transformer = (ast: any, options: generator_options) => {
           case 'WhileStatement':
           case 'ForStatement':
           case 'ForInStatement':
+          case 'AsyncFunctionDeclaration':
           case 'DoWhileStatement': {
             return replaceLoopStatement(x);
           }